Move client docs files to client.cms package
