+++ stream-chat-android-docs/src/main/kotlin/io/getstream/chat/docs/kotlin/client/cms/Messages.kt
@@ -1,4 +1,4 @@
-package io.getstream.chat.docs.kotlin
+package io.getstream.chat.docs.kotlin.client.cms
 
 import android.content.Context
 import io.getstream.chat.android.client.ChatClient
