+++ stream-chat-android-docs/src/main/kotlin/io/getstream/chat/docs/kotlin/client/cms/Push.kt
@@ -1,5 +1,5 @@
 @file:Suppress("DEPRECATION_ERROR")
-package io.getstream.chat.docs.kotlin
+package io.getstream.chat.docs.kotlin.client.cms
 
 import android.content.Context
 import com.google.firebase.messaging.FirebaseMessagingService
